push

+++ DiscordBotNet/LegendaryBot/Rarity.cs
@@ -1,6 +1,6 @@
 ﻿namespace DiscordBotNet.LegendaryBot;
 
-public enum Rarity : short
+public enum Rarity : byte
 {
     OneStar = 1, TwoStar, ThreeStar, FourStar, FiveStar
 }
